fix issue 140 wrong arrow angle
The steps to reproduce error:

var master = gitgraph.branch("master");
master.commit();
var branch0 = master.branch("branch0");
var branch1 = branch0.branch("branch1");
var branch2 = branch0.branch("branch2");

branch1.commit()
branch2.commit()

The arrow point from branch0 to branch2, branch3
will have vertical directions.

The cause:
When create a new commit object, and the commit is the first commit
to this branch, it will first check parentBranch has commit or not.
If it has commit, it will push a new point to path to create a joint
to the path, in order to create more beautiful path.
In the above case, the branch0 has no commit when branch1, 2
commit, so branch1, 2 will have no breakpoint in path, instead
the line will straightly draw from master to the commit.

The fix:
The check of parentCommit is empty or not is actually not needed.
We can always draw the joint.
The position of the joint is calculated as follow:
x equals to commit x, y equals to startpoint y minus template spacing

So if difference of y is larger than template spacing, it will create
a joint at one template spacing then straight line to new commit.

TODO:
Currently the angle of arrow is calculated by various condition.
It will check it is forkcommit or not, then determines that if
it is vertical or not.
I think a simpler way is to calculate angle of arrow by path.
First find the point in path that x, y equal to commit.
Then calculate angle by x,y difference of previous point and current
point