distinguish vertical case and horizontal case
When orientation setting to vertical and horizontal.
The position of joint in path is different.

The joint should be at one x/y spacing later to its parent node,
and at same y/x to child node.
Example, if the orientation is vertical, and we have to insert a
joint. The joint's x is same x to new commit, and y is a spacing
back to its parent.