Use branchToDisplay to compute branch path
The goal is to fix the temporal coupling between commits branches
initialisation and branches paths.

For the moment, we need to init branches paths when commit branches are
computed with first parent only. That makes things more complex.

If we manage to drop this temporal coupling, we'll be able to refactor
branches paths calculation a lot!