Change commit branches calculation
This remove temporal coupling between commit branches & branches paths.
It doesn't seem to have an impact on current rendering scenarios.

Only push "first parent" branch.

For example: commits of a branch won't be considered part of master.
That's because, for rendering, we consider commits of the branch being
at a different position than commits on "master".

But commits created on master before branch was created will be part
of both "master" and the branch.
=> might not be necessary as it's not consistent, but it works so far.